refactor(db): consolidate migrations into single initial schema
Consolidate 6 separate migrations into one comprehensive initial schema:

Migration Strategy:
- Remove incremental migrations (0cfaf48e5b9c, 314aef4924e4, 7610ccc63d69, 852ad2e01efe, b4301dc0075c, f67961867576)
- Create single migration c283e057a219_initial_schema_with_all_modules.py
- Includes complete schema for User, Competition, Enrollment, and Country modules

Database Schema:
- users table: Complete with country_code FK and all profile fields
- competitions table: Full schema with 3-country support and state management
- enrollments table: With custom_handicap and status tracking
- countries table: 198 countries with bilingual names (EN/ES)
- country_adjacencies table: 614 bidirectional border relationships

Seeds:
- alembic/seeds/__init__.py: Automated seeding of 198 countries + adjacencies
- Executed automatically on migration upgrade
- Data sourced from Wikipedia (accurate as of 2025)

Alembic Config:
- env.py: Updated imports for all modules' metadata
- Combined metadata from user, competition, and shared domains

Benefits:
- Cleaner migration history for new environments
- Single source of truth for initial schema
- Faster initial setup (one migration vs six)
- Automatic country data population
