+++ alembic/seeds/__init__.py
@@ -1,0 +1,1 @@
+# Alembic seeds module
