--- alembic/versions/852ad2e01efe_add_user_profile_fields.py
@@ -1,33 +1,0 @@
-"""add user profile fields
-
-Revision ID: 852ad2e01efe
-Revises: 314aef4924e4
-Create Date: 2025-11-12 17:32:04.609132
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '852ad2e01efe'
-down_revision: Union[str, None] = '314aef4924e4'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # No hay cambios estructurales en la base de datos
-    # UserIdDecorator es solo un tipo de SQLAlchemy, no afecta el schema
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # No hay cambios estructurales para revertir
-    pass
-    # ### end Alembic commands ###
