+++ alembic/versions/c283e057a219_initial_schema_with_all_modules.py
@@ -1,0 +1,140 @@
+"""initial_schema_with_all_modules
+
+Revision ID: c283e057a219
+Revises:
+Create Date: 2025-11-23 12:13:17.821158
+
+"""
+from typing import Sequence, Union
+import sys
+from pathlib import Path
+
+from alembic import op
+import sqlalchemy as sa
+
+# Importar seeds de países
+# Añadir el directorio seeds al path para poder importar
+seeds_dir = Path(__file__).parent.parent / 'seeds'
+sys.path.insert(0, str(seeds_dir))
+from seed_countries import seed_countries, seed_country_adjacencies
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'c283e057a219'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('countries',
+    sa.Column('code', sa.CHAR(length=2), nullable=False),
+    sa.Column('active', sa.Boolean(), nullable=False, server_default='true'),
+    sa.Column('name_en', sa.String(length=100), nullable=False),
+    sa.Column('name_es', sa.String(length=100), nullable=False),
+    sa.PrimaryKeyConstraint('code')
+    )
+    op.create_table('country_adjacencies',
+    sa.Column('country_code_1', sa.CHAR(length=2), nullable=False),
+    sa.Column('country_code_2', sa.CHAR(length=2), nullable=False),
+    sa.ForeignKeyConstraint(['country_code_1'], ['countries.code'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['country_code_2'], ['countries.code'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('country_code_1', 'country_code_2')
+    )
+    op.create_table('users',
+    sa.Column('id', sa.CHAR(length=36), nullable=False),
+    sa.Column('first_name', sa.String(length=50), nullable=False),
+    sa.Column('last_name', sa.String(length=50), nullable=False),
+    sa.Column('email', sa.String(length=255), nullable=False),
+    sa.Column('password', sa.String(length=255), nullable=False),
+    sa.Column('handicap', sa.Float(), nullable=True),
+    sa.Column('handicap_updated_at', sa.DateTime(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.Column('email_verified', sa.Boolean(), nullable=False, server_default='false'),
+    sa.Column('verification_token', sa.String(length=255), nullable=True),
+    sa.Column('country_code', sa.CHAR(length=2), nullable=True),
+    sa.ForeignKeyConstraint(['country_code'], ['countries.code'], ondelete='SET NULL'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email')
+    )
+    op.create_table('competitions',
+    sa.Column('id', sa.CHAR(length=36), nullable=False),
+    sa.Column('creator_id', sa.CHAR(length=36), nullable=False),
+    sa.Column('name', sa.String(length=200), nullable=False),
+    sa.Column('start_date', sa.Date(), nullable=False),
+    sa.Column('end_date', sa.Date(), nullable=False),
+    sa.Column('country_code', sa.CHAR(length=2), nullable=False),
+    sa.Column('secondary_country_code', sa.CHAR(length=2), nullable=True),
+    sa.Column('tertiary_country_code', sa.CHAR(length=2), nullable=True),
+    sa.Column('team_1_name', sa.String(length=100), nullable=False),
+    sa.Column('team_2_name', sa.String(length=100), nullable=False),
+    sa.Column('handicap_type', sa.String(length=20), nullable=False),
+    sa.Column('handicap_value', sa.Integer(), nullable=True),
+    sa.Column('max_players', sa.Integer(), nullable=False, server_default='24'),
+    sa.Column('team_assignment', sa.String(length=20), nullable=False, server_default="'MANUAL'"),
+    sa.Column('status', sa.String(length=20), nullable=False, server_default="'DRAFT'"),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['country_code'], ['countries.code'], ondelete='RESTRICT'),
+    sa.ForeignKeyConstraint(['creator_id'], ['users.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['secondary_country_code'], ['countries.code'], ondelete='RESTRICT'),
+    sa.ForeignKeyConstraint(['tertiary_country_code'], ['countries.code'], ondelete='RESTRICT'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('enrollments',
+    sa.Column('id', sa.CHAR(length=36), nullable=False),
+    sa.Column('competition_id', sa.CHAR(length=36), nullable=False),
+    sa.Column('user_id', sa.CHAR(length=36), nullable=False),
+    sa.Column('status', sa.String(length=20), nullable=False),
+    sa.Column('team_id', sa.String(length=10), nullable=True),
+    sa.Column('custom_handicap', sa.Numeric(precision=4, scale=1), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['competition_id'], ['competitions.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+
+    # Crear índices para optimizar queries
+    op.create_index('ix_users_country_code', 'users', ['country_code'], unique=False)
+    op.create_index('idx_verification_token', 'users', ['verification_token'], unique=False)
+    op.create_index('ix_competitions_creator_id', 'competitions', ['creator_id'], unique=False)
+    op.create_index('ix_competitions_status', 'competitions', ['status'], unique=False)
+    op.create_index('ix_competitions_start_date', 'competitions', ['start_date'], unique=False)
+    op.create_index('ix_enrollments_competition_id', 'enrollments', ['competition_id'], unique=False)
+    op.create_index('ix_enrollments_user_id', 'enrollments', ['user_id'], unique=False)
+    op.create_index('ix_enrollments_status', 'enrollments', ['status'], unique=False)
+
+    # Crear constraint única para enrollment (un usuario solo puede estar inscrito una vez por competición)
+    op.create_unique_constraint('uq_enrollment_competition_user', 'enrollments', ['competition_id', 'user_id'])
+
+    # Insertar datos de seed: 166 países + 614 relaciones de fronteras
+    seed_countries()
+    seed_country_adjacencies()
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    # Eliminar índices primero
+    op.drop_index('ix_enrollments_status', table_name='enrollments')
+    op.drop_index('ix_enrollments_user_id', table_name='enrollments')
+    op.drop_index('ix_enrollments_competition_id', table_name='enrollments')
+    op.drop_index('ix_competitions_start_date', table_name='competitions')
+    op.drop_index('ix_competitions_status', table_name='competitions')
+    op.drop_index('ix_competitions_creator_id', table_name='competitions')
+    op.drop_index('idx_verification_token', table_name='users')
+    op.drop_index('ix_users_country_code', table_name='users')
+
+    # Eliminar constraint única
+    op.drop_constraint('uq_enrollment_competition_user', 'enrollments', type_='unique')
+
+    # Eliminar tablas en orden inverso (respetando foreign keys)
+    op.drop_table('enrollments')
+    op.drop_table('competitions')
+    op.drop_table('users')
+    op.drop_table('country_adjacencies')
+    op.drop_table('countries')
+    # ### end Alembic commands ###
